Social: add batched /render-messages collection endpoint (SOCIAL-468)
Replaces the singular /render-message endpoint with a plural collection
endpoint. Accepts a post_id plus an items[] array (one entry per
connection) and returns one record per item — keyed by client-supplied
id — so the JS side can consume it via getEntityRecords.

Per-item failures are isolated as { id, error: { code, message } }, so
one bad render never fails the whole batch. Render-affecting inputs
({ network, message, is_social_post }) are deduped in-request: identical
inputs render once and fan out the result to every matching item.
